Change changeset-related functions to their own file.
